Added importer for stop times
Interestingly, stop times after midnight count as the same service day as the previous, so there are times like 25:22:00. Which means that Ruby's DateTime has a fit when trying to parse these. The work around is to store the hours, minutes and seconds as separate columns in the DB. We'll have to do some fun stuff elsewhere to accurately present this info.